Refactored the code to make it easier to work w/out creating merge conflicts
