--- src/actions.js
@@ -1,12 +1,0 @@
-/*
- * action types
- */
-export const EXAMPLE = 'EXAMPLE'
-
-
-/*
- * action creators
- */
-export function createExampleAction(parameter) {
-  return { type: EXAMPLE, parameter }
-}
