+++ src/actions/exampleAction.js
@@ -1,0 +1,18 @@
+// /*
+//  * action types
+//  */
+// export const EA1 = 'EXAMPLE_ACTION_1'
+// export const EA2 = 'EXAMPLE_ACTION_2'
+
+
+// /*
+//  * Optional: action creators (return an action object)
+//  */
+
+// export function createExampleAction1(parameter) {
+//   return { type: EA1, parameter }
+// }
+
+// export function createExampleAction2() {
+//   return { type: EA2 }
+// }
