--- src/reducers.js
@@ -1,4 +1,0 @@
-import { combineReducers } from 'redux'
-import {
-    EXAMPLE,
-  } from './actions'
