+++ src/reducers/exampleReducer.js
@@ -1,0 +1,15 @@
+// Define an initial state (if state is undefined, then it will be set to this)
+// initialState = 0
+
+// const exampleReducer = (state = initialState, action) => {
+//     switch(action.type) {
+//         case "EXAMPLE_ACTION_1":
+//             return {...state, variable: action.variable}
+//         case "EXAMPLE_ACTION_2":
+//             return {...state, variable: state.variable+1}
+//         default:
+//             // Return original state if the action type is unmatched
+//             return state
+//     }
+// }
+// export default exampleReducer;
