+++ src/reducers/index.js
@@ -1,0 +1,12 @@
+import {combineReducers} from 'redux'
+
+// 1. Import Reducers here
+// import exampleReducer from './exampleReducer'
+import emailpassReducer from './emailpassReducer'
+
+
+export default combineReducers({
+    // 2. Then add them here
+    // exampleReducer,
+    emailpassReducer,
+})
